refactor viewport panel
