+++ packages/editor/src/components/EditorContainer.tsx
@@ -31,7 +31,6 @@ import { HierarchyPanelTab } from '@ir-engine/ui/src/components/editor/panels/Hi
 import { MaterialsPanelTab } from '@ir-engine/ui/src/components/editor/panels/Materials'
 import { PropertiesPanelTab } from '@ir-engine/ui/src/components/editor/panels/Properties'
 import { ScenePanelTab } from '@ir-engine/ui/src/components/editor/panels/Scenes'
-import { ViewportPanelTab } from '@ir-engine/ui/src/components/editor/panels/Viewport'
 import { VisualScriptPanelTab } from '@ir-engine/ui/src/components/editor/panels/VisualScript'
 
 import ErrorDialog from '@ir-engine/ui/src/components/tailwind/ErrorDialog'
@@ -62,6 +61,7 @@ import { useTranslation } from 'react-i18next'
 import { IoHelpCircleOutline } from 'react-icons/io5'
 import { setCurrentEditorScene } from '../functions/sceneFunctions'
 import { FilesPanelTab } from '../panels/files'
+import { ViewportPanelTab } from '../panels/viewport'
 import './EditorContainer.css'
 
 export const DockContainer = ({ children, id = 'editor-dock', dividerAlpha = 0 }) => {
