+++ packages/editor/src/panels/viewport/tools/RenderTool.tsx
@@ -30,15 +30,15 @@ import { ShadowMapResolutionOptions } from '@ir-engine/client-core/src/user/comp
 import { useMutableState } from '@ir-engine/hyperflux'
 import { RendererState } from '@ir-engine/spatial/src/renderer/RendererState'
 import { RenderModes, RenderModesType } from '@ir-engine/spatial/src/renderer/constants/RenderModes'
+import BooleanInput from '@ir-engine/ui/src/components/editor/input/Boolean'
+import InputGroup from '@ir-engine/ui/src/components/editor/input/Group'
+import SelectInput from '@ir-engine/ui/src/components/editor/input/Select'
+import { Popup } from '@ir-engine/ui/src/components/tailwind/Popup'
+import Tooltip from '@ir-engine/ui/src/primitives/mui/Tooltip'
+import Button from '@ir-engine/ui/src/primitives/tailwind/Button'
 import { GiWireframeGlobe } from 'react-icons/gi'
 import { RiArrowDownSLine } from 'react-icons/ri'
 import { TbBallBowling, TbInnerShadowBottom, TbInnerShadowBottomFilled, TbShadow } from 'react-icons/tb'
-import Button from '../../../../../primitives/tailwind/Button'
-import Tooltip from '../../../../../primitives/tailwind/Tooltip'
-import { Popup } from '../../../../tailwind/Popup'
-import BooleanInput from '../../../input/Boolean'
-import InputGroup from '../../../input/Group'
-import SelectInput from '../../../input/Select'
 
 const renderModes: { name: RenderModesType; icon: JSX.Element }[] = [
   {
