+++ packages/editor/src/panels/viewport/tools/SceneHelpersTool.tsx
@@ -29,12 +29,12 @@ import { downloadScreenshot } from '@ir-engine/editor/src/functions/takeScreensh
 import { EditorHelperState, PlacementMode } from '@ir-engine/editor/src/services/EditorHelperState'
 import { useMutableState } from '@ir-engine/hyperflux'
 import { RendererState } from '@ir-engine/spatial/src/renderer/RendererState'
+import Button from '@ir-engine/ui/src/primitives/tailwind/Button'
+import Tooltip from '@ir-engine/ui/src/primitives/tailwind/Tooltip'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { LuMousePointerClick, LuMove3D } from 'react-icons/lu'
 import { twMerge } from 'tailwind-merge'
-import Button from '../../../../../primitives/tailwind/Button'
-import Tooltip from '../../../../../primitives/tailwind/Tooltip'
 import ColliderAtomsSvg from './icons/collider_atoms.svg?react'
 import RulerSvg from './icons/ruler.svg?react'
 import ScreenshotSvg from './icons/screenshot.svg?react'
